requirements.txt
praw==7.7.1
python-dotenv==1.0.0
pandas==2.1.4
gradio==4.44.0
colorama==0.4.6
argparse
datetime
json
csv
time
random
logging
os
sys